code_challenge_3/pyproject.toml
[project]
name = "code-challenge-3"
version = "0.1.0"
description = ""
authors = [
    {name = "Robert Tsui",email = "rt.roberttsui@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
